+++ NEWS.md
@@ -2,5 +2,11 @@
 
 * Initialise package and git repository (GitLab)
 
+# TROLL 0.1.9002
+
+* TROLLoutput class
+* print.TROLLoutput method
+* load.R function to open TROLL results in TROLLoutput class object
+